+++ http/src/main/scala/com/twitter/finatra/http/routing/HttpWarmup.scala
@@ -21,17 +21,22 @@ class HttpWarmup @Inject()(
   /* Public */
 
   /**
-   * Send a request to warmup services that are not yet externally receiving traffic
-   * @param request Request to send
-   * @param forceRouteToAdminHttpMuxers Route the request to the HttpMuxer (e.g. needed for twitter-server admin endpoints that do not start with /admin)
-   * @param responseCallback Callback called for every response where assertions can be made (note: be aware that failed assertions will prevent a server from
-   *                         restarting when dependent services are unresponsive)
+   * Send a request to warmup services that are not yet externally receiving traffic.
+   *
+   * @param request - [[com.twitter.finagle.http.Request]] to send
+   * @param forceRouteToAdminHttpMuxers - route the request to the HttpMuxer (e.g. needed for twitter-server admin
+   *                                    endpoints that do not start with /admin)
+   * @param times - number of times to send the request
+   * @param responseCallback - callback called for every response where assertions can be made. NOTE: be aware that
+   *                         failed assertions that throws Exceptions could prevent a server from restarting. This is
+   *                         generally when dependent services are unresponsive causing the warm-up request(s) to fail.
+   *                         As such, you should wrap your warm-up calls in these situations in a try/catch {}.
    */
   def send(
     request: => Request,
     forceRouteToAdminHttpMuxers: Boolean = false,
     times: Int = 1,
-    responseCallback: Response => Unit = identity) {
+    responseCallback: Response => Unit = identity _) {
 
     for (i <- 1 to times) {
       val response = executeRequest(request, forceRouteToAdminHttpMuxers)