+++ thrift/src/test/scala/com/twitter/finatra/thrift/tests/doeverything/DoEverythingThriftServer.scala
@@ -8,8 +8,6 @@ import com.twitter.finatra.thrift.routing.ThriftRouter
 import com.twitter.finatra.thrift.tests.doeverything.controllers.DoEverythingThriftController
 import com.twitter.finatra.thrift.tests.doeverything.filters.ExceptionTranslationFilter
 import com.twitter.finatra.thrift.{ThriftFilter, ThriftServer}
-import com.twitter.inject.Logging
-import com.twitter.inject.utils.Handler
 import com.twitter.util.NullMonitor
 
 object DoEverythingThriftServerMain extends DoEverythingThriftServer
@@ -46,13 +44,3 @@ class DoEverythingThriftServer extends ThriftServer {
   }
 }
 
-class DoEverythingThriftWarmupHandler
-  extends Handler
-  with Logging {
-
-  override def handle() = {
-    // TODO: implement ThriftWarmup
-    // warm up thrift service(s) here
-    info("Warm-up done.")
-  }
-}